Include arpa/inet.h instead of sys/endian.h
The header is included to get defintion for htons. In termux the
definition is in sys/endian.h, but on other systems it is in
arpa/inet.h.

arpa/inet.h includes sys/endian.h though, so including that header
works in termux as well.

Change to arpa/inet.h so that termux-api-package can be easily
compiled both in termux and for other systems.